requirements.txt
pandas
numpy
scikit-learn
tensorflow
joblib
schedule
clickhouse-connect
boto3
python-dotenv
scikeras
pyarrow
fastparquet